refactor: split database.rs into database/ module with separate hashmap implementations
